Added slurm scripts, plots, zero binning, corrections for parton showers

+++ examples/slurm_scripts/parton_shower_gen.sh
@@ -10,50 +10,94 @@
 ###################################
 # Preparation
 ###################################
+# Determining whether we use the syntax required for sbatch on the MIT supercloud
+supercloud_syntax=true
+
+# -------------------------
+# Help message:
+# -------------------------
+usage() {
+printf "###################################
+# Parton Shower Generation:
+###################################
+
+Options for $0:
+  # ============================
+  # Setting Parameters:
+  # ============================
+  [<--set_params|-t> <int>] :				Set params:
+  							0) Using the given inputs (see setup/set_params.sh);
+							1) With a particular choice of parameters for a non-perturbative shower:
+                                                          * Running coupling and MLL accuracy for observables;
+                                                          * Shower cutoff of LAMBDA_QCD ~ 340 MeV.
+							(see setup/set_np1.sh)
+  # ============================
+  # Other Options:
+  # ============================
+  [<--logfile|-l> ] :					Set log file;
+  [<--help|-h> ] :                                      Get this help message!\n\n"
+  1>&2; exit 1; }
+
+
+
 # ============================
-# Misc. Preparation
+# Setting Parameters:
 # ============================
-# -------------------------
-# Code Switches:
-# -------------------------
-# Physics parameters
-fixedcoup='False'
-jet_type='quark'
-obsacc='MLL'
+# Transform long options to short ones
+args=( )
+for arg; do
+    case "$arg" in
+	--set_params)		 args+=( -t ) ;;
+	--logfile)		 args+=( -l ) ;;
+        --help|-h)               usage ;;
+        *)                       args+=( "$arg" ) ;;
+    esac
+done
 
-# Parameters for event and function generation
-nevents='5e4'
-cutoff='1e-10'
-# Sets the shower cutoff for running coupling.
-# The usual choice is MU_NP. I also use 1e-8, 1e-10.
+# Extra parameters
+logfile='shower_logfile'
 
-load_events='False'
-save_events='False'
-save_correlations='True'
+while getopts "t:l:" OPTION; do
+    case $OPTION in
+    t)
+    case $OPTARG in
+      0) chmod +x setup/set_params.sh; ./setup/set_params.sh $@ ;;
+      1) chmod +x setup/set_params_NP1.sh; ./setup/set_params_NP1.sh ;;
+    esac
+    ;;
+    l) logfile=${OPTARG};;
+    esac
+done
 
-# -------------------------
-# Supercloud Preparation:
-# -------------------------
-# Preparation for running in supercloud cluster:
-module load anaconda/2021b
-
-# Linking log files to more precisely named logs
-logfile='shower_fc'$fixedcoup'_obs'$obsacc'_'$nevents'events_cutoff'$cutoff
-ln -f logs/zlog-${SLURM_JOB_ID}.out logs/$logfile.out.${SLURM_JOB_ID}
-ln -f logs/zlog-${SLURM_JOB_ID}.err logs/$logfile.err.${SLURM_JOB_ID}
 
 # ============================
 # Path preparation:
 # ============================
-# Should be run from the root folder /JetMonteCarlo
 
 # -------------------------
 # PYTHONPATH:
 # -------------------------
 # Adding the JetMonteCarlo directory to the PYTHONPATH
 # Must be used in the directory /path/to/JetMonteCarlo/
-chmod +x examples/slurm_scripts/prepare_path.sh
-./examples/slurm_scripts/prepare_path.sh
+chmod +x setup/prepare_path.sh
+./setup/prepare_path.sh
+
+# -------------------------
+# Log File Preparation:
+# -------------------------
+if [ supercloud_syntax = true ] ;
+then
+  # Preparation for running in supercloud cluster:
+  module load anaconda/2021b
+
+  # Linking slurm log files to more precisely named logs
+  ln -f logs/zlog-${SLURM_JOB_ID}.out logs/$logfile.out.${SLURM_JOB_ID}
+  ln -f logs/zlog-${SLURM_JOB_ID}.err logs/$logfile.err.${SLURM_JOB_ID}
+else
+  # Writing to log files without slurm
+  exec 1>logs/$logfile.out
+  exec 2>logs/$logfile.err
+fi
 
 
 ###################################
@@ -61,48 +105,31 @@ chmod +x examples/slurm_scripts/prepare_path.sh
 ###################################
 printf "# ============================
 # Date: "`date '+%F'`"-("`date '+%T'`")
-# ============================"
-python examples/params.py
+# ============================\n\n"
 
-# ============================
-# Setting desired accuracy:
-# ============================
-# Fixed coupling:
-sed -i "s/FIXED_COUPLING = .*/FIXED_COUPLING = "$fixedcoup"/" examples/params.py
-# Accuracy for f.c. observables and splitting functions is LL by default
-sed -i "s/OBS_ACC = .*/OBS_ACC = '"$obsacc"'/" examples/params.py
-
-# ============================
-# Setting jet type:
-# ============================
-sed -i "s/JET_TYPE = .*/JET_TYPE = '"$jet_type"'/" examples/params.py
-
-# ============================
-# Setting MC parameters:
-# ============================
-sed -i "s/NUM_SHOWER_EVENTS = .*/NUM_SHOWER_EVENTS = int("$nevents")/" examples/params.py
-sed -i "s/SHOWER_CUTOFF = .*/SHOWER_CUTOFF = "$cutoff"/" examples/params.py
-
-
-sed -i "s/LOAD_SHOWER_EVENTS = .*/LOAD_SHOWER_EVENTS = "$load_events"/" examples/event_generation/parton_shower_gen.py
-sed -i "s/SAVE_SHOWER_EVENTS = .*/SAVE_SHOWER_EVENTS = "$save_events"/" examples/event_generation/parton_shower_gen.py
-sed -i "s/SAVE_SHOWER_CORRELATIONS = .*/SAVE_SHOWER_CORRELATIONS = "$save_correlations"/" examples/event_generation/parton_shower_gen.py
+printf 'Running '"$0"' with options:'"\n"
+printf '%q ' "$@"
+printf "\n\n"
+set -- "${args[@]}"
 
 printf "
 # ============================
 # Parton shower generation:
 # ============================
 \n"
+# ============================
 # Generating shower samples and observables
-python examples/event_generation/parton_shower_gen.py
+# ============================
+# python examples/event_generation/parton_shower_gen.py
 
-
-# Remove duplicate log files:
-rm logs/zlog-${SLURM_JOB_ID}.out
-rm logs/zlog-${SLURM_JOB_ID}.err
+if [ supercloud_syntax = true ] ;
+then
+  # Remove duplicate log files:
+  rm logs/zlog-${SLURM_JOB_ID}.out
+  rm logs/zlog-${SLURM_JOB_ID}.err
+fi
 
 printf "# ============================
 # End time: "`date '+%F'`"-("`date '+%T'`")
 # ============================"
-
